+++ src/main/java/com/cfar/swim/worldwind/render/airspaces/ObstacleCylinder.java
@@ -76,10 +76,10 @@ public class ObstacleCylinder extends CappedCylinder implements Obstacle {
 	protected Depiction depiction = null;
 	
 	/** the threshold cost of this obstacle cylinder */
-	private int thresholdCost = 0;
+	private double thresholdCost = 0d;
 	
 	/** the active cost of this obstacle cylinder */
-	private int activeCost = 0;
+	private double activeCost = 0d;
 	
 	/**
 	 * Constructs an obstacle cylinder at the specified location with the
@@ -185,10 +185,10 @@ public class ObstacleCylinder extends CappedCylinder implements Obstacle {
 	 * 
 	 * @param threshold the threshold cost of this obstacle cylinder
 	 *
-	 * @see ThresholdRenderable#setThreshold(int)
+	 * @see ThresholdRenderable#setThreshold(double)
 	 */
 	@Override
-	public void setThreshold(int threshold) {
+	public void setThreshold(double threshold) {
 		this.thresholdCost = threshold;
 		this.updateVisibility();
 	}
@@ -201,7 +201,7 @@ public class ObstacleCylinder extends CappedCylinder implements Obstacle {
 	 * @see ThresholdRenderable#getThreshold()
 	 */
 	@Override
-	public int getThreshold() {
+	public double getThreshold() {
 		return this.thresholdCost;
 	}
 
@@ -326,7 +326,7 @@ public class ObstacleCylinder extends CappedCylinder implements Obstacle {
 		*/
 		
 		// TODO: rounding or conservative ceiling might be more appropriate
-		int cost = (this.costInterval.getCost() + other.costInterval.getCost()) / 2;
+		double cost = (this.costInterval.getCost() + other.costInterval.getCost()) / 2d;
 		
 		CostInterval costInterval = new CostInterval(this.costInterval.getId(), start, end, cost);
 		interpolant.setCostInterval(costInterval);
